# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-Dashboard API description.",
        "title": "Swagger E-Dashboard API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/edashboard",
    "paths": {
...
        "/lab/charges": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lab"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sRptType",
                        "name": "sRptType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "COMPANY_ID",
                        "name": "COMPANY_ID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/lab/kpi-statistics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lab"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sRptType",
                        "name": "sRptType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "COMPANY_ID",
                        "name": "COMPANY_ID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/lab/service-cls": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lab"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sRptType",
                        "name": "sRptType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "COMPANY_ID",
                        "name": "COMPANY_ID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/mortality/lookup/gender": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mortality"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/mortality/total-death": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mortality"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "sYear",
                        "name": "sYear",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/mortality/total-death-case": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mortality"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "sYear",
                        "name": "sYear",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/mortality/total-death-count-gender": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mortality"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "sYear",
                        "name": "sYear",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/mortality/total-death-gender": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mortality"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "sYear",
                        "name": "sYear",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sGender",
                        "name": "sGender",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}